foldRight not tail-recursive
Hello @winitzki  I am loving your book - very comprehensive and thorough introduction to folding - I found the material so compelling that I couldn't help putting it at the centre of part 2 of my slide deck on folding: https://www.slideshare.net/pjschwarz/folding-unfolded-polyglot-fp-for-fun-and-profit-haskell-and-scala-part-2 

But when I got to this point about the foldRight function in the Scala standard library not being tail recursive, I went to verify it and as far as I can tell it doesn't seem to be true. See slides 33 and 34 of the https://www.slideshare.net/pjschwarz/folding-unfolded-polyglot-fp-for-fun-and-profit-haskell-and-scala-part-2#33

Am I missing something? Please let me know.

Thanks.