Might it make sense to add foldRight to this table?
Hey Sergei,

I am fully aware that while this kind of suggestion is easy to make, it is much harder to implement (that's if you are at all convinced that it is worth doing), because it is not just a case of adding something, e.g. (1) there are preexisting reasons why it has not been done; e.g. (2) making the change would require other changes, and so on and so forth.

Still, allow me to share this half-baked idea with you.

See the PR for details.